+++ 3D BaseFramework/Src/Application/GameObject/UI/Player/HP/Player_HP.cpp
@@ -21,7 +21,7 @@ void Player_HP::Update()
 	else
 	{
 		m_NowDownTime = 0;
-		if (m_rect[HP::down].width != m_rect[HP::hp].width)m_rect[HP::down].width-=m_DownChange; //現在のHPまで徐々に減少
+		if (m_rect[HP::down].width != m_rect[HP::hp].width)m_rect[HP::down].width -= (long)m_DownChange; //現在のHPまで徐々に減少
 	}
 }
 
@@ -29,7 +29,7 @@ void Player_HP::DrawSprite()
 {
 	for (int h = 0; h < HP::num; ++h)
 	{
-		KdShaderManager::Instance().m_spriteShader.DrawTex(m_pTex[h], (int)m_pos.x, (int)m_pos.y, (int)m_rect[h].width * m_size, (int)m_rect[h].height * m_size, &m_rect[h], &m_color, m_pivot);
+		KdShaderManager::Instance().m_spriteShader.DrawTex(m_pTex[h], (int)m_pos.x, (int)m_pos.y, int(m_rect[h].width * m_size), int(m_rect[h].height * m_size), &m_rect[h], &m_color, m_pivot);
 	}
 }
